Moved box2d script files into separate folder
