+++ engine/gamesys/src/gamesys/wscript
@@ -20,7 +20,7 @@ def build(bld):
         includes = '. .. ../../src ../../proto',
         proto_gen_py = True,
         protoc_includes = '../../proto',
-        source = bld.path.ant_glob(['*.cpp', 'resources/*.cpp', 'components/*.cpp', 'scripts/*.cpp']) + bld.path.parent.parent.ant_glob('proto/gamesys/*'),
+        source = bld.path.ant_glob(['*.cpp', 'resources/*.cpp', 'components/*.cpp', 'scripts/*.cpp', 'scripts/box2d/*.cpp']) + bld.path.parent.parent.ant_glob('proto/gamesys/*'),
         target = 'gamesys')
 
     if bld.env.PLATFORM in ('arm64-ios', 'x86_64-ios'):
@@ -58,6 +58,8 @@ def build(bld):
         'scripts/script_image.cpp',
         'scripts/script_sys_gamesys.cpp',
         'scripts/script_http.cpp',
+        'scripts/box2d/script_box2d.cpp',
+        'scripts/box2d/script_box2d_body.cpp',
         'components/comp_sound.cpp',
         'components/comp_collection_proxy.cpp'] + proto_files)
